Create a batch service plugin
The batch plugin provides a mechanism for building and previewing sets
of proposed updates to multiple project/branches/refs that should be
applied in a batch. These updates are built with Gerrit changes.

This initial change provides a mechanism to merge a change to a batch
with a single command.

Change-Id: I3e470e8217457f4889abb64146750f87ad035f4b