Create a batch service plugin
The batch plugin provides a mechanism for building and previewing sets
of proposed updates to multiple project/branches/refs that should be
applied in a batch. These updates are built with Gerrit changes.

This initial change provides a mechanism to merge a change to a batch
with a single command.

Up-Stream url is the following, also mind the change-Ids are different:
https://gerrit-review.googlesource.com/c/plugins/batch/+/91439/2

Change-Id: I860fa82dfa6cfe2029c534c55593e580c6f8afc6