Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Maybe also implement bazel build too?

Whilst I like Bazel for Gerrit builds and core-plugins, I do think that raises the bar of adoptions of the plugins :-(

I remember people complaining on Twitter that extending Gerrit is difficult because of the pre-requisites needed to just build one.

Gerrit plugins should be *independent* from having a local Gerrit development environment.
People should be free to write plugins with the language they are familiar with and with the build system they know and they support.

Patch-set: 1
CC: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>