Update patch set 11
Patch Set 11:

(1 comment)

so, quite expectedly, this implements only a backend - a (sensible) frontend isn't possible as a plugin.

but this makes me wonder what the point of this really is? i mean, you can manage batches in refs hosted by gerrit without gerrit doing the management of the branch. our experience with "outsourcing" the batch management to gerrit is that it just adds a point of failure, so everything else being equal, i'd definitely avoid it.

for completeness, let me spam you (once more) with references to the qt project's implementation:
- https://codereview.qt-project.org/Documentation/cmd-staging-new-build.html (incl. pages linked at the bottom)
- https://codereview.qt-project.org/gitweb?p=qtqa/gerrit.git;a=commit;h=57cc39577e3de20fddc2b56f5a5e53bf6d33b890

Patch-set: 11
CC: Oswald Buddenhagen <1013267@173816e5-2b9a-37c3-8a2e-48639d4f1153>