Update patch set 11
Patch Set 11:

Hi Oswald,

> so, quite expectedly, this implements only a backend - a (sensible) frontend isn't possible as a plugin.

I'm not sure what a sensible frontend would provide, but why wouldn't it be possible as a plugin?

> 
> but this makes me wonder what the point of this really is? i mean, you can manage batches in refs hosted by gerrit without gerrit doing the management of the branch. our experience with "outsourcing" the batch management to gerrit is that it just adds a point of failure, so everything else being equal, i'd definitely avoid it.

You can do that, and in fact, we used to. We switched tactics and developed this plugin to address a few things:

1. Each client shares most of the implementation
We developed our 'ref-only' approach when we were part of one team, then we re-org'ed and found ourselves part of another team. The ref-only code was going to be hard to share between the two teams (and we envisioned/knew-of other teams coming later), so we wanted to move as much of the intelligence of that system into a shared plugin with simple APIs.

2. Related to (1), not all clients are responsible citizens
The 'ref-only' approach means clients do whatever they want with the refs, including leaving them around and not cleaning them up. For a busy project + CI system, that's many thousands of refs each week. The batch plugin includes the logic to automatically clean these up (after a configurable amount of time).

3. Reduced load
Pushing a new ref to Gerrit is (in our understanding) more expensive than using an API like the batch plugin provides. This is because we don't have to check any incoming objects and we entirely use server-side objects.

Martin might recall other reasons (he's on vacation till Monday), but those are the major ones I think.

As for adding a point of failure, aren't you fetching refs from Gerrit anyway? How is this adding another point?

> 
> for completeness, let me spam you (once more) with references to the qt project's implementation:
> - https://codereview.qt-project.org/Documentation/cmd-staging-new-build.html (incl. pages linked at the bottom)
> - https://codereview.qt-project.org/gitweb?p=qtqa/gerrit.git;a=commit;h=57cc39577e3de20fddc2b56f5a5e53bf6d33b890

Fun to see we're not the only ones still on 2.7... ;)
We do use this with our internal 2.7 Gerrit; however, we haven't cleaned up a 2.7 branch of batch because upstream stable-2.7 can't build this plugin (we added extensions, but didn't backport them).

I read through those links just now (must have missed them before, thanks for including again) and that's an interesting approach. In general, that looks like something you could build on top of this plugin for a specific workflow.

We also used Patch-Set-Lock and Approval-Lock (not yet upstreamed) approval functions to implement your STAGED and INTEGRATING change states. And we use the task plugin to provide UI feedback on change status wrt CI needs. It's possible that those differences alleviate the concerns/issues linked to in that gitweb doc you have above.

Let me know if there's anything else I can help explain.

Nasser

Patch-set: 11
Reviewer: Nasser Grainawi <1002666@173816e5-2b9a-37c3-8a2e-48639d4f1153>