Update patch set 11
Patch Set 11:

> for completeness, let me spam you (once more) with references to the qt project's implementation:
> - https://codereview.qt-project.org/Documentation/cmd-staging-new-build.html (incl. pages linked at the bottom)
> - https://codereview.qt-project.org/gitweb?p=qtqa/gerrit.git;a=commit;h=57cc39577e3de20fddc2b56f5a5e53bf6d33b890

Neat, thanks for posting these.  That seems like a neat implementation of a very similar idea.  Your implementation seems slightly higher level, but similar.  It seems focused a bit more on user workflows.  It is inspiring to see that we have effectively come up independently with some pretty similar ideas.  That gives me lots of confidence that this is actually the right direction, the natural evolution of the powerful code workflow engine that Gerrit is.  My suggestion above of a WUI (in #6) is something you seem to have done!

It was not clear to me from the rather small list of issues (which seemed rather like minor complaints specific to your implementation), why you think this approach does not work well.  Could you elaborate on what you see as problematic with such a system?  It would be trivial for your workflow system to be rebuilt with a plugin on top of the batch plugin.  That would be one way for you to use it on a more modern Gerrit if that is something you aspire to.

Patch-set: 11
Reviewer: Martin Fick <1003883@173816e5-2b9a-37c3-8a2e-48639d4f1153>