Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> Nasser, Martin, Nicholas:
> 
> This series has been up for review for nearly 2 years.  Are you waiting for agreement/consensus from the gerrit maintainers/users/other contributors before submitting it?
> 
> I would rather suggest that since this is your plugin, if it works for you you should just go ahead and submit it. I suspect that once it's submitted and available as a built jar file from CI, people will be more likely to try it.

Thanks Dave.  While we would certainly appreciate any external review, we are not waiting for it.  We have changed our approach a few times.  The first patchsets would only compile against our fork.  Then we ported the series to the Gerrit master.  Now we have decided to port things to Gerrit 2.15 stable so that it can be used on 2.15 and we will have follow up changes that port it to master.  I will merge the bulk of this series very soon once I am confident that it works well on 2.15.

Patch-set: 14
Reviewer: Gerrit User 1003883 <1003883@173816e5-2b9a-37c3-8a2e-48639d4f1153>